Do not use non portable ssize_t
The ssize_t type is not portable.  We have 2 avoidable uses of this
type.  In both instances, we cast the value returned by simm16_operand
(of type int16_t) to ssize_t before left-shifting it by 2.  The shift
operator automaticly implies an int promotion of the returned value to
int32_t.  However as we will add this to a uint64, we can change the
cast to int64_t directly.

Note that on Linux (64 bits), ssize_t to int64_t are equivalent, so this
patch is really a no-op.

While doing this modification, I also realized that some places did a
cast for similar reasons, but to std::ptrdiff_t.  As the casted offset
is really an offset in inferior memory and not in host memory, I also
convert those to use int64_t.

Note that similarly to our amd_dbgapi_global_address_t, we could add a
address difference type for inferior addresses:

    using amd_dbgapi_global_address_diff_t = int64_t;

and use this.  This might be done in a later patch.

+++ src/architecture.cpp
@@ -751,7 +751,7 @@ amdgcn_architecture_t::branch_target (wave_t &wave,
     {
       target
         = pc + instruction.size ()
-          + (static_cast<std::ptrdiff_t> (simm16_operand (instruction)) << 2);
+          + (static_cast<int64_t> (simm16_operand (instruction)) << 2);
     }
   else if (is_cbranch_g_fork (instruction))
     {
@@ -897,7 +897,7 @@ amdgcn_architecture_t::classify_instruction (
     {
       information.emplace_back (
         address + instruction.size ()
-        + (static_cast<std::ptrdiff_t> (simm16_operand (instruction)) << 2));
+        + (static_cast<int64_t> (simm16_operand (instruction)) << 2));
 
       if (sdst_regnum.has_value ())
         {
@@ -3953,7 +3953,7 @@ gfx10_architecture_t::branch_target (wave_t &wave,
       || is_subvector_loop_end (instruction))
     {
       return pc + instruction.size ()
-             + (static_cast<ssize_t> (simm16_operand (instruction)) << 2);
+             + (static_cast<int64_t> (simm16_operand (instruction)) << 2);
     }
 
   return gfx9_architecture_t::branch_target (wave, pc, instruction);
@@ -4342,7 +4342,7 @@ gfx10_architecture_t::classify_instruction (
         AMD_DBGAPI_INSTRUCTION_PROPERTY_NONE, instruction.size (),
         std::vector<uint64_t> (
           { address + instruction.size ()
-            + (static_cast<ssize_t> (simm16_operand (instruction)) << 2) })
+            + (static_cast<int64_t> (simm16_operand (instruction)) << 2) })
       };
     }
